+++ database/migrations/2024_03_21_194855_move_referral_percentage_column_from_settings_to_products_table.php
@@ -11,13 +11,13 @@ return new class extends Migration
      */
     public function up(): void
     {
-        if (Schema::hasColumn('settings', 'referral_percentage')) {
+        if (Schema::hasTable('settings') && Schema::hasColumn('settings', 'referral_percentage')) {
             Schema::table('settings', function (Blueprint $table) {
                 $table->dropColumn('referral_percentage');
             });
         }
 
-        if (!Schema::hasColumn('products', 'referral_percentage')) {
+        if (Schema::hasTable('products') && !Schema::hasColumn('products', 'referral_percentage')) {
             Schema::table('products', function (Blueprint $table) {
                 $table->double('referral_percentage')->nullable();
             });
